fastapi-gateway/pyproject.toml
[project]
name = "fastapi-gateway"
version = "0.1.0"
description = "FastAPI gateway proxying requests to Go gRPC micro-services."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "protobuf>=6.33.6",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.12.1",
    "pyrate-limiter>=4.0.2",
    "python-dotenv>=1.2.2",
    "fastapi[standard]>=0.129.0",
    "uvicorn>=0.41.0",
    "fastapi-limiter>=0.2.0",
    "httpx>=0.28.1",
    "redis>=5.0.0",
    "tenacity>=8.0.0",
    "minio>=7.2.18",
    "reportlab>=4.4.4",
    "sse-starlette>=3.4.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
